:sparkles: Move all sketches images to public folder on build
Next's Image component can only load images from inside the public folder. As it comes with a bunch of responsive goodies I thought it would be easier to do what it wants rather than hacking together a different image solution. I'll monitor performance of the move operation when we get to deployment.