:sparkles: Set up IDE
The babel setup for Next.js was causing a slowdown of sketch paints by ~6x compared to the webpack-dev-server that was running before. I couldn't cope with that speed when developing sketches locally, so this revives the non-Next setup just for local development. Still some kinks to work out with how this interacts with the routing in Next.