:memo: Replace README sketch images with links to Cloudinary
I tried to make a lambda which takes just the id, but didn't have access to the rest seeds in the README because getSketchContent relies on the sketches folder in the file system - this works fine on SSG pages but not API routes that happen at runtime. That method would also have had the issue of having to link to an absolute path (https://abstractpuzzl.es/api/...) and so rely on the deployment state, not the current branch. For now using a direct link to the Cloudinary image seems like the right solution.