:fire: Remove result assets from repo
All images are stored in Cloudinary (and backed up locally). We were using the canvas image in READMEs but that's now a direct link to Cloudinary. I was planning on using the pieces svg files to recreate the animation functionality on the current website but the file content can be generated at build time.