:bug: Replace references to document.createElement
In some sketches we create temporary canvases for layering and transforms, but on the server-side we can't reference document. This passes in a createCanvas function which uses document on the client side and node-canvas on the server side.

N.B. This is still causing errors on sketches that have multiple
canvases because c.getTransform isn't implemented in v2.7.0 if
node-canvas. This was fixed a week ago
(https://github.com/Automattic/node-canvas/pull/1769) but a new version
of the lib hasn't been released yet. I won't merge this until the
package upgrade.