:bug: Fix preview endpoint
Fix ternary and bleed scaling - this was an issue in sketches 045-050
because of how the canvas is copied and scaled, the slight difference
in sizes was multiplying to a big blur. There's still a bit of an
issue with scaling because the width/height in the design function is
inclusive of bleed so positioning within the canvas can differ
between the preview-generated image and the IDE. I think this is a
fundamental issue with how scaling works in this app and cba to do
anything about it.